% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists2df.R
\name{lists_to_df}
\alias{lists_to_df}
\title{Transforms list of records into a data.frame.}
\usage{
lists_to_df(l, list_columns = character(0), unnamed_default = NA_character_)
}
\arguments{
\item{l}{List of records. It is assumed, that all items with the same name share the same data type}

\item{list_columns}{Character vector with the names of the fields that should be treated as a nested lists.
They may contain heterogenous contents or simply be atomic types with length more than 1.}
}
\value{
Returns a data.table
}
\description{
Transforms list of records into a data.frame.
}
